network_operations: Perform BFS upto max_depth
BFS can be performed up to an optional max_depth. This ensures that the
entire tree is not searched beyond a max_depth (optionally). A test for
the same was also added, relying on the fact that the depth for any
vertex has a default value of INT_MAX when not processed by BFS.
Please note that DFS does not calculate the depth.